Netoyage des routes + Validator des données OK 100%

+++ app/Http/Controllers/FolderController.php
@@ -13,7 +13,10 @@ class FolderController extends Controller
 {
     public function OverView()
     {
-        $user_id = Auth::user()->id; // Récupérer l'utilisateur actuel
+        $user_id = Auth::user()->id;
+        if($user_id == null){
+            return view("home")->with("failure","La page des dossiers n'a pas pu être affiché");
+        }
         $root_folder_id = Folder::where("path","=","/files/user_" . $user_id)->first()->folder_id;
         return $this->View($root_folder_id);
     }
@@ -25,7 +28,7 @@ class FolderController extends Controller
         if ($folder) {
             return $folder->folder_id;
         }
-        return null;
+        return view("home")->with("failure","Une erreur s'est produite lors de l'obtention des dossiers");
     }
 
     public function getNoteIdFromPath($notePath) {
@@ -35,14 +38,19 @@ class FolderController extends Controller
         if ($note) {
             return $note->note_id;
         }
-        return null;
+        return view("home")->with("failure","Une erreur s'est produite lors de l'obtention des notes");
     }
 
     public function getFolderContents($folderId) {
 
         $folderPath = Folder::find($folderId)->path;
-            $directories = Storage::directories($folderPath);
-            $files = Storage::files($folderPath);
+
+        if(!$folderPath){
+            return redirect()->route("home")->with("failure","Une erreur s'est produite lors de l'obtention d'un chemin de dossier");
+        }
+
+        $directories = Storage::directories($folderPath);
+        $files = Storage::files($folderPath);
 
 
         $folderContents = [];
@@ -72,7 +80,19 @@ class FolderController extends Controller
 
     public function View(int $id)
     {
-        $folder_path = Folder::where("folder_id","=",$id)->first()->path;
+        $folder = Folder::where("folder_id","=",$id)->first();
+
+        if(!$folder){
+             return redirect()->route("home")->with("failure","Le dossier demandé n'est pas disponible");
+        }
+
+        $folder_path = $folder->path;
+
+        //dd($folder->owner_id . "vs" .Auth::user()->id);
+        // Check user authorization
+        if($folder->owner_id != Auth::user()->id){ // TODO système d'autorisation accès
+             return redirect()->route("home")->with("failure","Vous n'êtes pas autorisé à voir cette ressource");
+        }
 
         $folder_parent_path = dirname($folder_path);
 
@@ -92,88 +112,70 @@ class FolderController extends Controller
             }
 
         $folderContents = $this->getFolderContents($id);
+            //dd($folderContents);
         return view("folder.FolderView",
         ["folderContents" => $folderContents,
         "parent_content" => $parent_content,
-            "folder_path" => $folder_path]);
+            "folder_path" => $folder_path,
+            "folder_name" => $folder->name]);
     }
 
     public function Add()
     {
-        $user = Auth::user(); // Obtenez l'utilisateur authentifié
-        $folders = Folder::where('owner_id', $user->id)->get(); // Récupère les dossiers de l'utilisateur
+        $user_id = Auth::user()->id;
+        $folders = Folder::where('owner_id', $user_id)->get(); // Récupère les dossiers de l'utilisateur
         return view("folder.AddFolder",
             ['folders' => $folders]);
     }
 
     // POST
-    public function Store2(Request $request)
-    {
-        $user = Auth::user(); // Récupérer l'utilisateur actuel
-
-        // Validation des données du formulaire
-        $validatedData = $request->validate([
-            'folder_name' => 'required|max:100',
-            'location' => 'required|in:root,sub_folder', // S'assurer que la valeur est 'root' ou 'sub_folder'
-            'parent_folder_id' => 'nullable|exists:folders,folder_id' // Valider que l'ID du parent existe
-        ]);
-
-        // Création d'une nouvelle instance de Folder avec les données du formulaire
-        $newFolder = new Folder();
-        $newFolder->owner_id = $user->id;
-        //$newFolder->folder_name = $validatedData['folder_name'];
-
-        // Si l'utilisateur choisit d'ajouter à la racine
-        if ($validatedData['location'] === 'root') {
-            //$newFolder->parent_folder_id = null; // Laisser parent_folder_id comme NULL pour la racine
-            $newFolder->path = '/files/user_' . Auth::user()->id . "/" . $validatedData['folder_name'];
-
-        } else {
-            $path_parent = Folder::where("folder_id" , "=" , $validatedData["parent_folder_id"])->first()->path;
-            $newFolder->path = $path_parent . "/" . $validatedData["folder_name"];
-        }
-
-        // Enregistrement du nouveau dossier dans la base de données
-        Storage::makeDirectory($newFolder->path);
-        $newFolder->save();
-
-        // Redirection vers une autre page après la création du dossier (à ajuster selon vos besoins)
-        return redirect()->route('folder_overview')->with('success', 'Le dossier a été créer !');
-    }
 
 
     public  function Store(Request $request)
     {
-        // Nom du dossier à créer
-        $name = $request->get("add-dossier");
-        // Chemin du dossier à créer
-        $path_current = $request->get("path_current");
+        $validatedData = $request->validate([
+            'add-dossier' => ['required', 'regex:/^(?!.*[.]{2})[A-Za-z0-9]+(\.[A-Za-z0-9]+)?$/'],
+            'path_current' => 'required'
+        ]);
+        $name = $validatedData["add-dossier"];                  // Nom du dossier à créer
+        $path_current = $validatedData["path_current"];         // Chemin du dossier à créer
+        $user_id = Auth::user()->id;
 
         // Verif de la localisation
         if($path_current == "Racine")
-            $path_current = "/user_" . Auth::user()->id;
+            $path_current = "/user_" . $user_id;
         $path_current .= "/";
         $path_final = $path_current . $name;
-        //dd($path_final);
 
-        // Creation du model
-        $newFolder = new Folder();
-        $newFolder->owner_id = Auth::user()->id;
+
+        $newFolder = new Folder();                              // Creation du model
+        $newFolder->owner_id = $user_id;
         $newFolder->name = $name;
         $newFolder->path = $path_final;
 
-        // Persistance + save
-        Storage::makeDirectory($newFolder->path);
+
+        $check = Storage::makeDirectory($newFolder->path);               // Persistance + save
+
+        if(!$check){
+            return redirect()->route("home")->with("failure","Une erreur s'est produite lors de la création d'un dossier");
+        }
         $newFolder->save();
         return redirect()->back()->with("success","Le dossier a bien été créer !");
     }
 
     public function Delete(Request $request)
     {
-        $id =  $request->get("id");
+        $validatedData = $request->validate([
+           'id' => ["integer"]
+        ]);
+        $id =  $validatedData["id"];
         $folder = Folder::find($id);
         //dd(storage_path($folder->path));
         //dd(Storage::exists($folder->path));
+
+        if(!$folder)
+            return redirect()->route("home")->with("failure","Une erreur s'est produite lors de la recherche d'un dossier");
+
         Storage::deleteDirectory($folder->path);
         $folder->delete();
         return redirect()->back()->with(["success" => "Dossier supprimé avec succès"]);
